Remove old files, add new flock file, updates to fcntl_lock
Remove the old files which contained the old code for the demo.

Now fcntl and flock locks are applied each one on a different program
with a different source to keep things quite simple.

Add a case to fcntl_lock.c to request an exclusive lock on a file
previusly locked with a shared lock.

Some minor changes and bug fixes.